+++ VanControlCenter/LCDStringList.cpp
@@ -27,7 +27,7 @@ void LCDStringList::addElement(String elem){
 
 void LCDStringList::clear(){
 	//Replace all char except \n with with space
-	for (int i = 0; i < buffer.length(); i++){
+	for (uint16_t i = 0; i < buffer.length(); i++){
 		if (buffer.charAt(i) != '\n'){
 			buffer.setCharAt(i, ' ');
 		}
